+++ src/backend/admin/features/adminPermission/SaveServerConfigs.php
@@ -13,8 +13,8 @@ class SaveServerConfigs extends HasAdminPermission {
 	/**
 	 * @throws PageFlowException
 	 */
-	private function extractServerName(stdClass $obj): string {
-		$serverName = urldecode($obj->serverName);
+	private function extractServerName(array $obj): string {
+		$serverName = urldecode($obj['serverName']);
 		$len = strlen($serverName);
 		if($len < 3 || $len > 30)
 			throw new PageFlowException("The server name '$serverName' needs to consist of 3 to 30 characters");
@@ -24,39 +24,70 @@ class SaveServerConfigs extends HasAdminPermission {
 			return $serverName;
 	}
 	function exec(): array {
-		$serverStore = Configs::getDataStore()->getServerStore();
-		
-		if(!($settingsCollection = json_decode(Main::getRawPostInput())))
+		if(!($settingsCollection = json_decode(Main::getRawPostInput(), true)))
 			throw new PageFlowException('Unexpected data');
 		
-		if(!isset($settingsCollection->_))
-			throw new PageFlowException('No default settings');
+		if(!isset($settingsCollection['configs']) || !isset($settingsCollection['translationData']))
+			throw new PageFlowException('Missing data');
 		
+		$serverStore = Configs::getDataStore()->getServerStore();
 		$oldLangCodes = Configs::get('langCodes');
 		
+		
+		
 		$serverNames = [];
-		$langCodes = [];
-		foreach($settingsCollection as $code => $obj) {
-			if($code !== '_') {
-				array_push($langCodes, $code);
-				if(($k = array_search($code, $oldLangCodes)) !== false)
-					unset($oldLangCodes[$k]);
-			}
+		foreach($settingsCollection['translationData'] as $code => $obj) {
+			if(($k = array_search($code, $oldLangCodes)) !== false)
+				unset($oldLangCodes[$k]);
 			
 			$serverNames[$code] = $this->extractServerName($obj);
 			
-			$impressum = urldecode($obj->impressum);
+			$impressum = urldecode($obj['impressum']);
 			if(strlen($impressum))
 				$serverStore->saveImpressum($impressum, $code);
 			else
 				$serverStore->deleteImpressum($code);
 			
-			$privacyPolicy = urldecode($obj->privacyPolicy);
+			$privacyPolicy = urldecode($obj['privacyPolicy']);
 			if(strlen($privacyPolicy))
 				$serverStore->savePrivacyPolicy($privacyPolicy, $code);
 			else
 				$serverStore->deletePrivacyPolicy($code);
 		}
+		
+		$settingsCollection['configs']['serverName'] = $serverNames;
+		FileSystemBasics::writeServerConfigs($settingsCollection['configs']);
+		
+		
+		
+//		if(!isset($settingsCollection->_))
+//			throw new PageFlowException('No default settings');
+//
+//		$oldLangCodes = Configs::get('langCodes');
+//
+//		$serverNames = [];
+//		$langCodes = [];
+//		foreach($settingsCollection as $code => $obj) {
+//			if($code !== '_') {
+//				$langCodes[] = $code;
+//				if(($k = array_search($code, $oldLangCodes)) !== false)
+//					unset($oldLangCodes[$k]);
+//			}
+//
+//			$serverNames[$code] = $this->extractServerName($obj);
+//
+//			$impressum = urldecode($obj->impressum);
+//			if(strlen($impressum))
+//				$serverStore->saveImpressum($impressum, $code);
+//			else
+//				$serverStore->deleteImpressum($code);
+//
+//			$privacyPolicy = urldecode($obj->privacyPolicy);
+//			if(strlen($privacyPolicy))
+//				$serverStore->savePrivacyPolicy($privacyPolicy, $code);
+//			else
+//				$serverStore->deletePrivacyPolicy($code);
+//		}
 		
 		//if a language has been removed, we need to remove its files too:
 		foreach($oldLangCodes as $code) {
@@ -64,10 +95,10 @@ class SaveServerConfigs extends HasAdminPermission {
 			$serverStore->deletePrivacyPolicy($code);
 		}
 		
-		FileSystemBasics::writeServerConfigs([
-			'serverName' => $serverNames,
-			'langCodes' => $langCodes,
-		]);
+//		FileSystemBasics::writeServerConfigs([
+//			'serverName' => $serverNames,
+//			'langCodes' => $langCodes,
+//		]);
 		
 		return [];
 	}