+++ src/backend/config/configs.default.php
@@ -20,6 +20,7 @@ return [
 	
 	'csv_delimiter' => ';',
 	
-	'serverName' => ['_' => ''], //_ is the default language. All alternative languages need to be added via language-code (langCodes needs to be updates as well)
-	'langCodes' => []
+	'serverName' => ['en' => ''], //_ is the default language. All alternative languages need to be added via language-code (langCodes needs to be updates as well)
+	'defaultLang' => 'en',
+	'langCodes' => [] //will be used for privacyPolicy, legal and servername. The optional languages of the UI are hardcoded and will not be influenced by this value
 ];