+++ src/frontend/js/variables/defaults.js
@@ -265,14 +265,3 @@ Defaults.account = {
 	write: [],
 	publish: []
 };
-
-Defaults.serverSettings = {
-	langCodes: [],
-	$: {
-		translated: {
-			impressum: "",
-			serverName: "",
-			privacyPolicy: "",
-		}
-	}
-}