+++ src/frontend/pages/settings.js
@@ -8,10 +8,22 @@ import {Admin} from "../js/main_classes/admin";
 import {OwnMapping} from "../js/helpers/knockout_own_mapping";
 import {Defaults} from "../js/variables/defaults";
 import {PromiseCache} from "../js/main_classes/promise_cache";
-import {add_lang} from "../js/shared/lang_configs";
+import {addLang, changeLang} from "../js/shared/lang_configs";
 import {DetectChange} from "../js/main_classes/detect_change";
 import {NavigationRow} from "../js/main_classes/navigation_row";
 import {Requests} from "../js/main_classes/requests";
+
+let defaultServerSettings = {
+	langCodes: ["en"],
+	defaultLang: "en",
+	$: {
+		translated: {
+			impressum: "",
+			serverName: "",
+			privacyPolicy: "",
+		}
+	}
+}
 
 export function ViewModel(page) {
 	let self = this;
@@ -19,39 +31,62 @@ export function ViewModel(page) {
 	page.title(Lang.get("server_settings"));
 	this.promiseBundle = [
 		Admin.init(page),
-		PromiseCache.loadJson(FILE_ADMIN+"?type=get_serverConfigs", function(data) {
-			let serverSettings = OwnMapping.fromJS(Defaults.serverSettings, Defaults.serverSettings);
+		PromiseCache.loadJson(FILE_ADMIN+"?type=get_serverConfigs", function({configs, translationData}) {
+			let serverSettings = OwnMapping.fromJS(configs, defaultServerSettings);
+			let detector = new DetectChange(serverSettings);
+			let translations = {__detector: detector};
 			
-			serverSettings.serverName(data.serverName._);
-			serverSettings.impressum(data.impressum._);
-			serverSettings.privacyPolicy(data.privacyPolicy._);
-			
-			let langCodes = data.langCodes;
-			for(let i=langCodes.length-1; i>=0; --i) {
-				let code = langCodes[i];
-				let langObj = {};
-				langObj.serverName = data.serverName.hasOwnProperty(code) ? data.serverName[code] : '';
-				langObj.impressum = data.impressum.hasOwnProperty(code) ? data.impressum[code] : '';
-				langObj.privacyPolicy = data.privacyPolicy.hasOwnProperty(code) ? data.privacyPolicy[code] : '';
+			for(let langCode in translationData) {
+				let translationsSet = translationData[langCode];
+				for(let key in translationsSet) {
+					configs[key] = translationsSet[key];
+				}
 				
-				OwnMapping.add_lang(serverSettings, Defaults.serverSettings, langObj, code);
-				serverSettings.langCodes.push(ko.observable(code));
+				let lang = OwnMapping.bindNewLanguageContainer(serverSettings, configs);
+				translations[langCode] = lang;
+				detector.addMonitored(lang);
 			}
 			
-			let detector = new DetectChange(serverSettings);
-			detector.setDirty(false);
-			
-			return [serverSettings, detector];
+			// detector.addMonitored(translations);
+			return [serverSettings, translations, detector];
 		})
 	];
-	this.extraContent = "<lang-chooser params='enableAdd: true, alwaysVisible: true, hasTitle: true, langObj: $root.dataObj, langDefaults: $root.defaults'></lang-chooser>";
-	this.dataObj = OwnMapping.fromJS(Defaults.serverSettings, Defaults.serverSettings);
-	this.defaults = Defaults.serverSettings;
+	this.extraContent = "<lang-chooser params='hasTitle: true, langObj: $root.dataObj, translations: $root.translations'></lang-chooser>";
+	this.dataObj = null;
 	
 	this.isPreRelease = ko.observable(0);
 	this.releases = ko.observableArray();
 	this.downloadUrl = null
 	this.markdownRenderer = new MarkdownIt();
+	
+	let versionIsBelowThen =  function(newVersionString) {
+		let integersOld = PACKAGE_VERSION.match(/(\d+)\.(\d+)\.(\d+)\D*(\d*)/);
+		let integersNew = newVersionString.match(/(\d+)\.(\d+)\.(\d+)\D*(\d*)/);
+		
+		return integersOld && integersNew &&
+			(
+				integersNew[1] > integersOld[1] // e.g. 2.0.0 > 1.0.0
+				|| (
+					integersNew[1] === integersOld[1]
+					&& (
+						integersNew[2] > integersOld[2] // e.g. 2.1.0 > 2.0.0
+						|| (
+							integersNew[2] === integersOld[2]
+							&& (
+								integersNew[3] > integersOld[3] // e.g. 2.1.1 > 2.1.0
+								|| (
+									integersNew[3] === integersOld[3]
+									&& (
+										(integersOld[4] !== '' && integersNew[4] === '') // e.g. 2.1.1 > 2.1.1-alpha.1
+										|| (integersOld[4] !== '' && integersNew[4] !== '' && integersNew[4] > integersOld[4]) // e.g. 2.1.1-alpha.2 > 2.1.1-alpha.1
+									)
+								)
+							)
+						)
+					)
+				)
+			);
+	}
 	
 	let checkForUpdate = function() {
 		page.loader.showLoader(Lang.get("state_loading"), PromiseCache.loadText(URL_RELEASES_LIST, function(jsonString) {
@@ -62,11 +97,13 @@ export function ViewModel(page) {
 			let searchingForRelease = true;
 			for(let i=releases.length-1; i>=0; --i) {
 				let {tag_name, prerelease, body, published_at, assets} = releases[i];
-				if(searchingForRelease) {
-					if(tag_name.match(PACKAGE_VERSION+"$"))
-						searchingForRelease = false;
-					continue;
-				}
+				if(!versionIsBelowThen(tag_name))
+					continue
+				// if(searchingForRelease) {
+				// 	if(tag_name.match(PACKAGE_VERSION+"$"))
+				// 		searchingForRelease = false;
+				// 	continue;
+				// }
 				
 				let data = {version: tag_name, date: new Date(published_at), changeLog: body, downloadUrl: assets[0].browser_download_url};
 				if(prerelease)
@@ -91,12 +128,13 @@ export function ViewModel(page) {
 	
 	this.isPreRelease.subscribe(checkForUpdate);
 	
-	this.preInit = function(index, admin, [serverSettings, detector]) {
-		this.dataObj = serverSettings;
-		this.add_lang = add_lang.bind(this, serverSettings, Defaults.serverSettings);
+	this.preInit = function(index, admin, [serverSettings, translations, detector]) {
+		// changeLang(serverSettings, translations, serverSettings.defaultLang());
+		self.translations = translations;
+		self.dataObj = serverSettings;
 		NavigationRow.admin.change_observed(
 			detector,
-			self.changeText
+			self.saveServerSettings
 		);
 	};
 	this.postInit = checkForUpdate;
@@ -109,10 +147,12 @@ export function ViewModel(page) {
 	this.privacyPolicy = ko.observable();
 	this.currentVersion = PACKAGE_VERSION;
 	
+	
+	this.translations = null;
+	
 	this.hasUpdate = ko.observable(false);
 	this.noConnectionToUpdate = ko.observable(false);
 	this.newVersion = ko.observable("");
-	this.changelog = ko.observable("");
 	
 	
 	this.remove_lang = function(code) {
@@ -120,44 +160,40 @@ export function ViewModel(page) {
 		self.dataObj.langCodes.splice(index, 1);
 	}
 	
-	this.changeText = function() {
-		let getValues = function(code) {
-			let obj = OwnMapping.toLangJs(self.dataObj, code);
-			if(obj.impressum)
-				obj.impressum = encodeURIComponent(obj.impressum);
-			if(obj.privacyPolicy)
-				obj.privacyPolicy = encodeURIComponent(obj.privacyPolicy);
-			
-			if(!obj.serverName || obj.serverName.length < 3 || obj.serverName.length > 30) {
-				page.loader.info(Lang.get("error_short_serverName", "en"));
-				return false;
-			} else if(!check_string(obj.serverName)) {
-				page.loader.info(Lang.get("error_forbidden_characters"));
-				return false;
-			}
-			return obj;
-		};
-
-		let s = getValues("_");
-		if(!s)
-			return;
-		let settings = {_: s};
-
+	this.saveServerSettings = function() {
+		
+		
+		
+		let translationData = {};
 		let langCodes = self.dataObj.langCodes();
 		for(let i=langCodes.length-1; i>=0; --i) {
 			let code = langCodes[i]();
-			let currentObj = getValues(code);
-			if(!currentObj)
-				return;
-			currentObj.lang = code;
-			settings[code] = currentObj;
+			changeLang(self.dataObj, self.translations, code);
+			translationData[code] = {
+				serverName: self.dataObj.serverName(),
+				impressum: self.dataObj.impressum(),
+				privacyPolicy: self.dataObj.privacyPolicy(),
+			};
 		}
+		console.log({
+			configs: {
+				langCodes: OwnMapping.toJS(langCodes),
+				defaultLang: self.dataObj.defaultLang()
+			},
+			translationData: translationData
+		})
 
 		return page.loader.loadRequest(
 			FILE_ADMIN + "?type=save_serverConfigs",
 			false,
 			"post",
-			JSON.stringify(settings)
+			JSON.stringify({
+				configs: {
+					langCodes: OwnMapping.toJS(langCodes),
+					defaultLang: self.dataObj.defaultLang()
+				},
+				translationData: translationData
+			})
 		);
 	};